# speed up any invocation of counterfeiter